# internal/openapi/swagger.yaml
basePath: /api
definitions:
...
  api.APIErrorEnvelope:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: true
        type: boolean
      message:
        example: invalid request
        type: string
    type: object
...
  api.ManagedPeerDTO:
    properties:
      description:
        example: My Phone
        type: string
      dns:
        example: 8.8.8.8
        type: string
      enabled:
        example: true
        type: boolean
      presharedKey:
        example: JJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJJ=
        type: string
      privateKey:
        example: IIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIII=
        type: string
      publicKey:
        example: HHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHH=
        type: string
      tunnelIP:
        example: 10.10.0.2
        type: string
    type: object
...
  api.ManagedServerBackupDTO:
    properties:
      address:
        example: 10.10.0.1
        type: string
      description:
        example: My VPN
        type: string
      dns:
        example: 8.8.8.8
        type: string
      endpoint:
        example: vpn.example.com:51821
        type: string
      i1:
        type: string
      i2:
        type: string
      interfaceName:
        example: Wireguard1
        type: string
      listenPort:
        example: 51821
        type: integer
      mask:
        example: 255.255.255.0
        type: string
      mtu:
        example: 1420
        type: integer
      natEnabled:
        example: true
        type: boolean
      peers:
        items:
          $ref: '#/definitions/api.ManagedPeerDTO'
        type: array
      policy:
        example: none
        type: string
      privateKey:
        example: oA...
        type: string
    type: object
  api.ManagedServerBackupFile:
    properties:
      exportedAt:
        type: string
      managedServers:
        items:
          $ref: '#/definitions/api.ManagedServerBackupDTO'
        type: array
      type:
        type: string
      version:
        type: integer
    type: object
...
  api.ManagedServerDriftEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.ManagedServerDriftResponse'
      success:
        example: true
        type: boolean
    type: object
  api.ManagedServerDriftResponse:
    properties:
      drift:
        items:
          $ref: '#/definitions/api.ManagedServerBackupDTO'
        type: array
    type: object
  api.ManagedServerExportEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.ManagedServerBackupFile'
      success:
        example: true
        type: boolean
    type: object
  api.ManagedServerImportEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.ManagedServerRestoreResponse'
      success:
        example: true
        type: boolean
    type: object
  api.ManagedServerImportRequest:
    properties:
      managedServers:
        items:
          $ref: '#/definitions/api.ManagedServerBackupDTO'
        type: array
      options:
        $ref: '#/definitions/api.RestoreOptionsDTO'
      type:
        type: string
      version:
        type: integer
    type: object
...
  api.ManagedServerRestoreDriftRequest:
    properties:
      options:
        $ref: '#/definitions/api.RestoreOptionsDTO'
    type: object
  api.ManagedServerRestoreResponse:
    properties:
      outcomes:
        items:
          $ref: '#/definitions/api.RestoreOutcomeDTO'
        type: array
    type: object
...
  api.RestoreOptionsDTO:
    properties:
      allowRenumber:
        example: false
        type: boolean
    type: object
  api.RestoreOutcomeDTO:
    properties:
      action:
        example: created
        type: string
      addedPeers:
        example: 2
        type: integer
      conflicts:
        items:
          type: string
        type: array
      error:
        type: string
      name:
        example: Wireguard1
        type: string
      newName:
        example: Wireguard2
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /managed/drift:
    get:
      description: Returns settings.json entries whose NDMS interface is absent.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ManagedServerDriftEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: List managed servers missing from NDMS
      tags:
      - managed
  /managed/export:
    get:
      description: Returns a JSON backup file with every managed server including
        private keys.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ManagedServerExportEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Export all managed servers
      tags:
      - managed
  /managed/import:
    post:
      consumes:
      - application/json
      description: Restores managed servers from a backup file. Per-server atomic
        with pre-flight conflict detection.
      parameters:
      - description: backup contents + options
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ManagedServerImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ManagedServerImportEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Import a managed-server backup
      tags:
      - managed
  /managed/restore-drift:
    post:
      consumes:
      - application/json
      description: Detects drift internally then runs Restore on it. Convenience entry
        for the boot-time banner.
      parameters:
      - description: options
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.ManagedServerRestoreDriftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ManagedServerImportEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Restore drifted managed servers
      tags:
      - managed
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"